💄 Design : 탈퇴 버튼 수정 페이지로 이동 #54 #57
